requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.0.3
psycopg2==2.9.1
PyJWT==2.1.0
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
